code refactor plus issue #54 and issue #55
