+++ src/HatTrick.Text.Templating/Tag/Tag.cs
@@ -11,6 +11,7 @@ namespace HatTrick.Text.Templating
         #region internals
         private TagKind _kind;
         private string _tag;
+        private int _tagLength;
         private TrimMark _markers;
         private bool _forceTrim;
         #endregion
@@ -23,7 +24,8 @@ namespace HatTrick.Text.Templating
         public Tag(string tag, bool forceTrim)
         {
             _tag = tag;
-            _kind = TagKind.Unknown;
+            _tagLength = tag.Length;
+            _kind = TagKind.Simple;
             _markers = TrimMark.None;
             _forceTrim = forceTrim;
             this.Init();
@@ -42,50 +44,38 @@ namespace HatTrick.Text.Templating
         private void ResolveKind()
         {
             string tag = _tag;
-            if (Tag.IsIfTag(tag))                               //# if logic tag (boolean switch)
-            {
+            if (Tag.IsIfTag(tag))                       //# if logic tag (boolean switch)
                 _kind = TagKind.If;
-            }
-            else if (Tag.IsEndIfTag(tag))                       //end if
-            {
+
+            else if (Tag.IsEndIfTag(tag))               //end if
                 _kind = TagKind.EndIf;
-            }
-            else if (Tag.IsEachTag(tag))                        //#each enumeration
-            {
+
+            else if (Tag.IsEachTag(tag))                //#each enumeration
                 _kind = TagKind.Each;
-            }
-            else if (Tag.IsEndEachTag(tag))                     //end each
-            {
+
+            else if (Tag.IsEndEachTag(tag))             //end each
                 _kind = TagKind.EndEach;
-            }
-            else if (Tag.IsWithTag(tag))                        //#with tag
-            {
+
+            else if (Tag.IsWithTag(tag))                //#with tag
                 _kind = TagKind.With;
-            }
-            else if (Tag.IsEndWithTag(tag))                     //end with
-            {
+
+            else if (Tag.IsEndWithTag(tag))             //end with
                 _kind = TagKind.EndWith;
-            }
-            else if (Tag.IsPartialTag(tag))                     //sub template tag
-            {
+
+            else if (Tag.IsPartialTag(tag))             //sub template tag
                 _kind = TagKind.Partial;
-            }
-            else if (Tag.IsVariableDeclareTag(tag))             //variable declaratino 
-            {
+
+            else if (Tag.IsVariableDeclareTag(tag))     //variable declaratino 
                 _kind = TagKind.VarDeclare;
-            }
-            else if (Tag.IsVariableAssignTag(tag))              //variable assignment
-            {
+
+            else if (Tag.IsVariableAssignTag(tag))      //variable assignment
                 _kind = TagKind.VarAssign;
-            }
-            else if (Tag.IsCommentTag(tag))                     //comment tag
-            {
+
+            else if (Tag.IsCommentTag(tag))             //comment tag
                 _kind = TagKind.Comment;
-            }
-            else                                                //simple tag
-            {
+
+            else                                        //simple tag
                 _kind = TagKind.Simple;
-            }
         }
         #endregion
 
@@ -96,22 +86,17 @@ namespace HatTrick.Text.Templating
                 return;
 
             if (_tag[1] == '-')                     //has discard left trim mark...
-            {
                 _markers = TrimMark.DiscardLeft;
-            }
+
             else if (_tag[1] == '+')                //has retain left trim mark...
-            {
                 _markers = TrimMark.RetainLeft;
-            }
+
 
             if (_tag[_tag.Length - 2] == '-')       //has discard right trim mark...
-            {
                 _markers |= TrimMark.DiscardRight;
-            }
+
             else if (_tag[_tag.Length - 2] == '+')   //has retain right trim mark...
-            {
                 _markers |= TrimMark.RetainRight;
-            }
         }
         #endregion
 
@@ -193,7 +178,9 @@ namespace HatTrick.Text.Templating
         public static bool IsCommentTag(string tag)
         {
             return tag[0] == '{'
-                && (tag[1] == '!' || (tag[1] == '-' && tag[2] == '!') || (tag[1] == '+' && tag[2] == '!'));
+                && (tag[1] == '!' || (tag[1] == '-' 
+                && tag[2] == '!') || (tag[1] == '+' 
+                && tag[2] == '!'));
         }
         #endregion
 
@@ -201,7 +188,9 @@ namespace HatTrick.Text.Templating
         public static bool IsPartialTag(string tag)
         {
             return tag[0] == '{'
-                && (tag[1] == '>' || (tag[1] == '-' && tag[2] == '>') || (tag[1] == '+' && tag[2] == '>'));
+                && (tag[1] == '>' || (tag[1] == '-' 
+                && tag[2] == '>') || (tag[1] == '+' 
+                && tag[2] == '>'));
         }
         #endregion
 
@@ -231,111 +220,65 @@ namespace HatTrick.Text.Templating
 		#region bind as
 		public string BindAs()
         {
-            char[] tag = _tag.ToCharArray();
             string bindAs = null;
             TagKind kind = _kind;
+
+            int start = 0;
+            int len = 0;
+            int maxLen = 0;
+
+            bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
+            bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
+
             switch (kind)
             {
                 case TagKind.Simple:
-                    {
-                        bindAs = new string(tag, 1, tag.Length - 2);
-                    }
+                    start = 1;
+                    maxLen = _tagLength - 2;
                     break;
                 case TagKind.If:
-                    {
-                        bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-                        bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-                        int start = left ? 5 : 4;
-
-                        int len = (left && right)
-                            ? (tag.Length - 7)
-                            : (left || right)
-                                ? (tag.Length - 6)
-                                : (tag.Length - 5);
-
-                        bindAs = new string(tag, start, len);
-                    }
+                    start = left ? 5 : 4;
+                    maxLen = 7;
                     break;
                 case TagKind.Each:
-                    {
-                        bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-                        bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-                        int start = left ? 7 : 6;
-
-                        int len = (left && right)
-                            ? (tag.Length - 9)
-                            : (left || right)
-                                ? (tag.Length - 8)
-                                : (tag.Length - 7);
-
-                        bindAs = new string(tag, start, len);
-                    }
+                    start = left ? 7 : 6;
+                    maxLen = 9;
                     break;
                 case TagKind.VarDeclare:
-                    {
-						bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-						bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-						int start = left ? 6 : 5;
-
-						int len = (left && right)
-							? (tag.Length - 8)
-							: (left || right)
-								? (tag.Length - 7)
-								: (tag.Length - 6);
-
-                        bindAs = new string(tag, start, len);
-					}
+					start = left ? 6 : 5;
+                    maxLen = 8;
                     break;
                 case TagKind.VarAssign:
-                    {
-						bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-						bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-						int start = left ? 3 : 2;
-
-						int len = (left && right)
-							? (tag.Length - 5)
-							: (left || right)
-								? (tag.Length - 4)
-								: (tag.Length - 3);
-
-                        bindAs = new string(tag, start, len);
-					}
+					start = left ? 3 : 2;
+                    maxLen = 5;
                     break;
                 case TagKind.With:
-                    {
-                        bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-                        bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-                        int start = left ? 7 : 6;
-
-                        int len = (left && right)
-                            ? (tag.Length - 9)
-                            : (left || right)
-                                ? (tag.Length - 8)
-                                : (tag.Length - 7);
-
-                        bindAs = new string(tag, start, len);
-                    }
+                    start = left ? 7 : 6;
+                    maxLen = 9;
                     break;
                 case TagKind.Partial:
-                    {
-                        bool left = this.HasTrimMark(TrimMark.DiscardLeft) || this.HasTrimMark(TrimMark.RetainLeft);
-                        bool right = this.HasTrimMark(TrimMark.DiscardRight) || this.HasTrimMark(TrimMark.RetainRight);
-                        int start = left ? 3 : 2;
-
-                        int len = (left && right) 
-                            ? (tag.Length - 5)
-                            : (left || right)
-                                ? (tag.Length - 4)
-                                : (tag.Length - 3);
-
-                        bindAs = new string(tag, start, len);
-                    }
+                    start = left ? 3 : 2;
+                    maxLen = 5;
                     break;
                 case TagKind.Comment:
-                case TagKind.Unknown:
-                    bindAs = null;
-                    break;
+                    throw new MergeException($"encountered un-expected TagKind: {kind} ... Comment tags cannot be bound");
             }
+
+            if (kind == TagKind.Simple) //simple tags cannot have trim markers...
+                len = maxLen;
+
+            else if (left && right)
+                len = (_tagLength - maxLen);
+
+            else if (left || right)
+                len = _tagLength - (maxLen - 1);
+
+            else
+                len = _tagLength - (maxLen - 2);
+
+
+			bindAs = (len > 0) ? new string(_tag.ToCharArray(), start, len) : null;
+
             return bindAs;
         }
         #endregion
@@ -343,7 +286,8 @@ namespace HatTrick.Text.Templating
         #region has trim mark
         public bool HasTrimMark(TrimMark marker)
         {
-            return (_markers & marker) == marker; 
+            bool exists = (_markers & marker) == marker;
+            return exists;
         }
         #endregion
 
@@ -361,13 +305,10 @@ namespace HatTrick.Text.Templating
             bool result = HasTrimMark(TrimMark.DiscardRight) || (_forceTrim && !HasTrimMark(TrimMark.RetainRight));
             return result;
         }
-		#endregion
+        #endregion
 
-		#region to string
-		public override string ToString()
-        {
-            return _tag;
-        }
+        #region to string
+        public override string ToString() => _tag;
         #endregion
     }
 }
