+++ src/HatTrick.Text.Templating/Tag/TagKind.cs
@@ -5,7 +5,6 @@ namespace HatTrick.Text.Templating
     #region tag kind enum
     public enum TagKind
     {
-        Unknown,
         Simple,
         If,
         EndIf,
