+++ src/HatTrick.Text.Templating/Tag/TrimMark.cs
