+++ test/sample-templates/literal-variable-declarations-in.txt
@@ -10,14 +10,28 @@ salutation: {:salutation}
 
 
 
-******** Declaration and Usage of Numeric Literals ********
+******** Declaration and Usage of Positive Numeric Literals ********
 {?var:uno = 1i }		{! int32 }
-{?var:dos = 2.0d }		{! double }
+{?var:dos = +2.0d }		{! double }
 {?var:tres = 3.000m }	{! decimal }
 
 uno + dos + tres = {(:uno, :dos, :tres) => sumIntDoubleDecimal}	{! pass literal declared variables to func +}
 uno + dos + tres = {(1, 2.00, 3.00) => sumIntDoubleDecimal}		{! pass literals directly to func +}
 
+******** Declaration and Usage of Unsigned Numeric Literals ********
+{?var:negUno = -1i }		{! int32 }
+{?var:negDos = -2.0d }		{! double }
+{?var:negTres = -3.000m }	{! decimal }
+
+******** Declaration and Usage of Signed Numeric Literals ********
+uno + dos + tres = {(:negUno, :negDos, :negTres) => sumIntDoubleDecimal}		{! pass literal declared variables to func +}
+uno + dos + tres = {(-1, -2.00, -3.00) => sumIntDoubleDecimal}					{! pass literals directly to func +}
+uno + dos + tres = {(+1, +2.00, +3.00) => sumIntDoubleDecimal}					{! pass literals directly to func +}
+uno + dos + tres = {(-1, -.200, -.300) => sumIntDoubleDecimal}					{! pass literals directly to func +}
+
+******** Sign followed by decimal **********
+{?var: negPercent = -.25m }
+negative percent: {:negPercent}
 
 
 ******** Declaration and Usage of String Literals Containing Quotes ********
